Add `GetMeadowSerialPortsForWindows()`.
Functioning more like the macOS and Linux counterparts, `GetMeadowSerialPortsForWindows()` runs a `WMIC` query to find Meadow ports on the machine. It specifically looks in the `Win32_PnPEntity` object for a device that has the class `Ports`, and a device ID that has Wilderness Labs' USB VID of 2E6A.

This allows the skipping of the false positives that are returned through the original `SerialPort.GetPortNames()` call. However, that functionality has been kept as a fallback, in case the WMIC query fails or otherwise returns no results.

An implementation note: I used running an external process of `wmic.exe`. This is not ideal, because it does lead to needing to parse the stdout, which could change as windows does. However, the System.Management classes are not available in .NET Standard 2.0, so as long as this Library targets that framework, this might be the closest we can get.