`GetSerialPortsForWindows()` with `System.Management` WMI.
As per discussion in #111, this is `GetSerialPortsForWindows()` using instead the .NET System.Management APIs. This introduces a dependency on the (Microsoft owned/provided) package `System.Management`, and implicitly `System.CodeDom`.

However, it is substantially faster than the invocation of an external executable and result parsing demonstrated in #111. It still provides the benefit of filtering the available ports to specifically find USB devices with Wilderness Labs' VID, which significantly reduces potential friction for any users with more than a single COM port (USB or otherwise.)