Create package with / path separator always
Windows `CreateEntryFromFile` will use the `\` windows path separator character when writing the zip file entry.  This might be fine and technically within spec, but when the update is extracted on the device it does not create the correct folder structure (so it seems `app\file.ext` is considered a file instead of a file within the `app` folder).

This just ensures on windows, the `/` separator replaces `\` instances when creating the package to avoid the issue.